Update fa-support-plugin.php
We need to save the BLOG ID to the cookie because we can't use get_current_blog_id, we need to know which blog they came from and if they are on current blog then we won't know if they came from an agent site. This is when a form is subbmited on main site.